Adds a feature flag to control whether the client cache function is just a passthrough. before we land breaking chagnes for the next major it will be off and then we can flag it on when we want to break it.
